#DCMEEREQ-69 Added Time Zone Support for Retrieve
Time Zone support is now fully supported.
Notes:
require integration tests
-Currently if no date is specified with a query today's date  is used
and time zone will not take care of daylight saving.

-Currently date time range is working and considers daylight saving
however in the case where the first part of the range exceeds that of
the second part due to daylight saving the SQL query is not adjusted to
that yet.