Created BPE replacement with dynamic decorator as a prototype
-Added a configuration for composition(container) and composable(entry
in a map) based on calling AET on the ArchiveAEExt
-added the ldap schema changes,
*To create a sample config create a container with
objectclass=dcmComposition then add a node with objectclass
dcmCollection under it and for each composable add a
objectclass=dcmComposable under the collection.
-Created a dynamic decorator that is loaded by default in the
descriptor, the decorator loads the appropriate implementation to invoke
for the decorated method based on the configuration which can be either
a pojo or an ejb, this simulates pipelines based on callingAET and can
be based on source (file, scp etc..( by changing the configuration.

if more methods are decorated in the dynamic decorator the method has to
be added to the BasicPipeLine interface since all retrieved
implementations are cast to BasicPipeLine.