[#ARCH-169] Provide File Alias Table
A join table is implemented on instance entity and a bidirectional
relationship to file ref to provide other linked file refs through the
join table, the same relation as before was left unchanged.

The retrieve query was adjusted to accommodate the changes and include
the new left joins with the new file refs provided through the file
alias table, their file system and time zone.