fix: [DCMEEREQ-799] Ensure Wado-URI and Wado-RS always decompresses (possibly) faulty JPEG-LS Lossless data
This only works if the configured JPEG-LS decoder can decompress faulty JPEG-LS correctly.
This is the case if ISO2JAI_IF_APP_OR_COM or no patching is configured in
ImageReaderFactory.properties for JAI-ImageIO or Agfa/Pegasus is used for decompression
(which automatically handles both faulty and correct JPEG-LS correctly).

Note: Wado-RS bulkdata request not yet handled in this commit (see TODOs in WadoRS).